Added AnnotationImpex interface.
Added annotation impex for JSON and Spark.
Broke out types to own file.
