+++ src/main/scala/org/broadinstitute/hail/driver/ImportAnnotations.scala
@@ -109,13 +109,13 @@ object ImportAnnotationsJSON extends Command {
 
     val t = Parser.parseType(options.`type`)
 
-    val extractVariant = Annotation.jsonExtractVariant(t, options.variantFields)
+    val extractVariant = JSONAnnotationImpex.jsonExtractVariant(t, options.variantFields)
 
     val rdd =
       sc.union(files.map { f =>
         sc.textFile(f)
           .map { line =>
-            Annotation.fromJson(parse(line), t, "<root>")
+            JSONAnnotationImpex.importAnnotation(parse(line), t, "<root>")
           }
       })
         .flatMap { va =>
