+++ src/test/scala/org/broadinstitute/hail/methods/ExprSuite.scala
@@ -12,28 +12,28 @@ import org.testng.annotations.Test
 class ExprSuite extends SparkSuite {
 
   @Test def exprTest() {
-    val symTab = Map("i" ->(0, TInt),
-      "j" ->(1, TInt),
-      "d" ->(2, TDouble),
-      "d2" ->(3, TDouble),
-      "s" ->(4, TString),
-      "s2" ->(5, TString),
-      "a" ->(6, TArray(TInt)),
-      "m" ->(7, TInt),
-      "as" ->(8, TArray(TStruct(("a", TInt),
+    val symTab = Map("i" -> (0, TInt),
+      "j" -> (1, TInt),
+      "d" -> (2, TDouble),
+      "d2" -> (3, TDouble),
+      "s" -> (4, TString),
+      "s2" -> (5, TString),
+      "a" -> (6, TArray(TInt)),
+      "m" -> (7, TInt),
+      "as" -> (8, TArray(TStruct(("a", TInt),
         ("b", TString)))),
-      "gs" ->(9, TStruct(("noCall", TGenotype),
+      "gs" -> (9, TStruct(("noCall", TGenotype),
         ("homRef", TGenotype),
         ("het", TGenotype),
         ("homVar", TGenotype),
         ("hetNonRef35", TGenotype))),
-      "t" ->(10, TBoolean),
-      "f" ->(11, TBoolean),
-      "mb" ->(12, TBoolean),
-      "is" ->(13, TString),
-      "iset" ->(14, TSet(TInt)),
-      "genedict" ->(15, TDict(TInt)),
-      "structArray" ->(16, TArray(TStruct(
+      "t" -> (10, TBoolean),
+      "f" -> (11, TBoolean),
+      "mb" -> (12, TBoolean),
+      "is" -> (13, TString),
+      "iset" -> (14, TSet(TInt)),
+      "genedict" -> (15, TDict(TInt)),
+      "structArray" -> (16, TArray(TStruct(
         ("f1", TInt),
         ("f2", TString),
         ("f3", TInt)))))
@@ -280,8 +280,8 @@ class ExprSuite extends SparkSuite {
   @Test def testJSON() {
     check(forAll { (t: Type) =>
       val a = t.genValue.sample()
-      val json = t.makeJSON(a)
-      a == Annotation.fromJson(json, t, "")
+      val json = t.toJSON(a)
+      a == JSONAnnotationImpex.importAnnotation(json, t)
     })
   }
 
@@ -290,7 +290,8 @@ class ExprSuite extends SparkSuite {
       val sb = new StringBuilder
       t.pretty(sb, 0)
       val a = t.genValue.sample()
-      t.makeSparkReadable(t.makeSparkWritable(a)) == a
+      JSONAnnotationImpex.importAnnotation(
+        JSONAnnotationImpex.exportAnnotation(a, t), t) == a
     })
   }
